New_Structure/requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
scikit-learn
statsmodels
plotly
pygwalker
langchain-google-genai
langchain-community
langchain
faiss-cpu
python-dotenv
